Index V2 consumption (#4462)
## Change
This change adds the ability to actually use the V2 index database. This
primarily required fixing the ignored search functionality from the
previous PR and updating the `ISource` implementation to be aware of V2
and how to retrieve the package version manifest files. Additionally, a
few places in the code were updated to avoid requesting all of the
package versions when we could instead use only the latest version. This
prevents the need to retrieve the package version file at all, which is
the goal for all read-only operations (`search`, `list`, `update` [aka
`list` but filtered to things with updates]). In reality, we do need to
get the package version files when we need to do ARP version mapping,
but that constitutes only 5% of the total packages in `winget-pkgs`
currently.

While the V2 usage code is always enabled, an experimental feature
(`indexV2`) is required to attempt to retrieve the V2 index from the
CDN. Disabling this feature should naturally transition one back to V1
on the next update attempt.

File caching, based on the known SHA256 hash of the file, is added as a
general utility. This is then used for V1 manifest files and also for V2
package version data and manifest files. While the files are generally
small, it is not uncommon for them to be used a few times over a short
period.

Thanks to @msftrubengu for pointing out that I forgot the include the
exports in the .def file in the previous PR. 🥇

+++ src/AppInstallerCLITests/TestCommon.cpp
@@ -87,16 +87,19 @@ namespace TestCommon
 
     TempFile::~TempFile() try
     {
-        switch (s_TempFileDestructorBehavior)
+        if (m_destructionToken)
         {
-        case TempFileDestructionBehavior::Delete:
-            std::filesystem::remove_all(_filepath);
-            break;
-        case TempFileDestructionBehavior::Keep:
-            break;
-        case TempFileDestructionBehavior::ShellExecuteOnFailure:
-            s_TempFilesOnFile.emplace_back(std::move(_filepath));
-            break;
+            switch (s_TempFileDestructorBehavior)
+            {
+            case TempFileDestructionBehavior::Delete:
+                std::filesystem::remove_all(_filepath);
+                break;
+            case TempFileDestructionBehavior::Keep:
+                break;
+            case TempFileDestructionBehavior::ShellExecuteOnFailure:
+                s_TempFilesOnFile.emplace_back(std::move(_filepath));
+                break;
+            }
         }
     }
     CATCH_LOG_RETURN()
@@ -105,6 +108,11 @@ namespace TestCommon
     {
         std::filesystem::rename(GetPath(), newFilePath);
         _filepath = newFilePath;
+    }
+
+    void TempFile::Release()
+    {
+        m_destructionToken = false;
     }
 
     void TempFile::SetDestructorBehavior(TempFileDestructionBehavior behavior)
